feat(core): add in/out edges to ComputedNode
