+++ packages/core/src/compute-view/EdgeBuilder.ts
@@ -1,4 +1,4 @@
-import { equals, head, keys, pluck, reject } from 'rambdax'
+import { keys, pluck } from 'rambdax'
 import type { Fqn, Relation } from '../types'
 import type { ComputedEdge, EdgeId } from '../types/computed-view'
 import { commonAncestor } from '../utils/fqn'
@@ -24,7 +24,7 @@ export class EdgeBuilder {
         const targets = this._relationsObj[source] ?? {}
         return keys(targets).map(target => {
           const relations = (targets[target] ?? []).sort(compareRelations)
-          const label = head(reject(equals(''), pluck('title', relations))) ?? null
+          const label = relations.find(r => r.title !== '')?.title ?? null
           return {
             id: `${source}:${target}` as EdgeId,
             parent: commonAncestor(source, target),
